requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2.2
django-import-export==4.3.7
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
